CB-1992 refactor base web platform (#824)
* CB-1992 refactor base web platform

* CB-1992 review fixes

* CB-1992 add copyright
