+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/server/CBConstants.java
@@ -63,7 +63,6 @@ public class CBConstants {
 
     public static final String ENV_CB_HOME = "CLOUDBEAVER_HOME";
 
-    public static final String DEFAUL_APP_ANONYMOUS_ROLE_NAME = "user";
 
     public static final String VAR_CB_LOCAL_HOST_ADDR = "CB_LOCAL_HOST_ADDR";
     public static final String VAR_HOST_DOCKER_INTERNAL = "host.docker.internal";
