+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/server/CBPlatformActivator.java
@@ -1,3 +1,5 @@
+package io.cloudbeaver.server;
+
 /*
  * DBeaver - Universal Database Manager
  * Copyright (C) 2010-2022 DBeaver Corp and others
@@ -14,31 +16,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudbeaver.server;
-
-import org.eclipse.core.resources.IWorkspace;
-import org.jkiss.code.NotNull;
-import org.jkiss.dbeaver.model.app.DBPPlatform;
-import org.jkiss.dbeaver.registry.BaseWorkspaceImpl;
 
 /**
- * Web workspace.
- *
- * Basically just a wrapper around Eclipse workspace.
+ * The activator class controls the plug-in life cycle
  */
-public class CBWorkspace extends BaseWorkspaceImpl {
+public class CBPlatformActivator extends WebPlatformActivator {
 
-    private final String workspaceId;
-
-    public CBWorkspace(DBPPlatform platform, IWorkspace eclipseWorkspace) {
-        super(platform, eclipseWorkspace);
-        workspaceId = readWorkspaceId();
-    }
-
-    @NotNull
     @Override
-    public String getWorkspaceId() {
-        return workspaceId;
+    protected void shutdownPlatform() {
+        try {
+            // Dispose core
+            if (CBPlatform.instance != null) {
+                CBPlatform.instance.dispose();
+            }
+        } catch (Throwable e) {
+            e.printStackTrace();
+            System.err.println("Internal error after shutdown process:" + e.getMessage()); //$NON-NLS-1$
+        }
     }
-
 }
